The GAT layers has been uploaded
The pyGAT repository is taken as a basis (https://github.com/Diego999/pyGAT), it is a rewritten version of the official implementation of the authors of the GAT paper. This implementation seemed to me the best of the existing ones in the implementation due to the way the structure of the layer is arranged.